simple types are not to referenced anymore when wrapped
Reason: A simple type like a string was aliased by the simple type name
but when it was wrapped in a complex type sequence, a pointer to the
alias was required. Since there is nothing like pointers to immutable
types, this means that declaring fields to be a pointer to a simple time
works, but there is no way to initialize the field. 1. The constants
generated for simple sequence types are no pointers (how could they be),
2. it's impossible to create such a simple type pinter because you can't
get the address of a string/int/u name it. This is hopefully fixed by
recording such type aliases during template processing.

Known issue: the anytype xsd type and the time/date/datetime types could
be pointers but are now treated as simple alias types.